feat: Complete AI system refactoring with proactive automation
Major architectural improvements to justify $800/month Workspace Pro pricing:

## Backend Refactoring (Go Idiomatic)
- Moved AI business logic from services/ to internal/ai/ following Go best practices
- Implemented priority-based task queue using container/heap
- Created 6 specialized processors for different AI tasks
- Added sophisticated analysis engines for issues and PRs
- Integrated WebSocket support for real-time updates

## AI Processors
- IssueProcessor: Intelligent triage with sentiment analysis and categorization
- PRProcessor: Comprehensive review with risk assessment and auto-approval
- ReportProcessor: Daily repository health reports
- StaleProcessor: Automated management of inactive items
- SecurityProcessor: Pattern-based vulnerability scanning
- DependencyProcessor: Dependency update tracking

## Frontend Enhancements
- Beautiful AI dashboard with real-time status cards
- Interactive automation controls with toggle switches
- Live activity feed with auto-refresh
- Configurable AI behavior settings (conservative/balanced/aggressive)
- Modern UI with animations and gradients

## Event-Driven Features
- Proactive response to repository events
- Configurable response delays
- Priority-based task processing
- Automatic retry with exponential backoff
- Context-based cancellation for graceful shutdown

This completes the AI platform to compete with Claude CLI and Gemini CLI,
providing enterprise-grade automation that justifies premium pricing.

+++ services/actions.go
@@ -209,7 +209,7 @@ func (e *ActionExecutor) executeJob(job *ActionJob) error {
 		status = "failed"
 	}
 	models.LogActivity("action_executed", fmt.Sprintf("Action %s %s", action.Title, status),
-		fmt.Sprintf("Action %s %s after %.1f seconds", action.Title, status, run.Duration),
+		fmt.Sprintf("Action %s %s after %.1f seconds", action.Title, status, float64(run.Duration)),
 		"system", action.RepoID, "action", action.ID)
 	
 	return execErr
